+++ android/src/main/java/com/jiguang/jpush/JPushPlugin.java
@@ -1,8 +1,12 @@
 package com.jiguang.jpush;
 
+import android.app.NotificationChannel;
+import android.app.NotificationManager;
 import android.content.BroadcastReceiver;
 import android.content.Context;
 import android.content.Intent;
+import android.net.Uri;
+import android.os.Build;
 import android.os.Bundle;
 import android.os.Handler;
 import android.os.Looper;
@@ -34,44 +38,30 @@ import io.flutter.plugin.common.MethodChannel.Result;
  * JPushPlugin
  */
 public class JPushPlugin implements FlutterPlugin, MethodCallHandler {
-
-
     private static String TAG = "| JPUSH | Flutter | Android | ";
-
-    public static JPushPlugin instance;
-
-    static List<Map<String, Object>> openNotificationCache = new ArrayList<>();
-
-    private boolean dartIsReady = false;
-    private boolean jpushDidinit = false;
-
-    private List<Result> getRidCache;
-
     private Context context;
-    private MethodChannel channel;
-    public Map<Integer, Result> callbackMap;
     private int sequence;
-
     public JPushPlugin() {
-        this.callbackMap = new HashMap<>();
         this.sequence = 0;
-        this.getRidCache = new ArrayList<>();
-        instance = this;
     }
-
 
     @Override
     public void onAttachedToEngine(FlutterPluginBinding flutterPluginBinding) {
-        channel = new MethodChannel(flutterPluginBinding.getBinaryMessenger(), "jpush");
+        MethodChannel  channel = new MethodChannel(flutterPluginBinding.getBinaryMessenger(), "jpush");
         channel.setMethodCallHandler(this);
         context = flutterPluginBinding.getApplicationContext();
+        JPushHelper.getInstance().setMethodChannel(channel);
+        JPushHelper.getInstance().setContext(context);
     }
 
 
     @Override
     public void onDetachedFromEngine(FlutterPluginBinding binding) {
-        channel.setMethodCallHandler(null);
-        instance.dartIsReady = false;
+        MethodChannel  channel =JPushHelper.getInstance().getChannel();
+        if (channel!=null) {
+            channel.setMethodCallHandler(null);
+        }
+        JPushHelper.getInstance().setDartIsReady(false);
     }
 
     @Override
@@ -123,11 +113,60 @@ public class JPushPlugin implements FlutterPlugin, MethodCallHandler {
             setAuth(call, result);
         } else if (call.method.equals("testCountryCode")) {
             testCountryCode(call, result);
-        }else if (call.method.equals("enableAutoWakeup")) {
+        } else if (call.method.equals("enableAutoWakeup")) {
             enableAutoWakeup(call, result);
+        } else if (call.method.equals("setLbsEnable")) {
+            setLbsEnable(call, result);
+        } else if (call.method.equals("setChannelAndSound")) {
+            setChannelAndSound(call, result);
         } else {
             result.notImplemented();
         }
+    }
+    public void setChannelAndSound(MethodCall call, Result result) {
+        HashMap<String, Object> readableMap = call.arguments();
+        if (readableMap == null) {
+            result.success(false);
+            return;
+        }
+        String channel = (String)readableMap.get("channel");
+        String channelId = (String)readableMap.get("channel_id");
+        String sound = (String)readableMap.get("sound");
+        try {
+            NotificationManager manager= (NotificationManager) context.getSystemService("notification");
+            if(Build.VERSION.SDK_INT<26){
+                result.success(false);
+                return;
+            }
+            if(TextUtils.isEmpty(channel)||TextUtils.isEmpty(channelId)){
+                result.success(false);
+                return;
+            }
+            NotificationChannel channel1=new NotificationChannel(channelId,channel, NotificationManager.IMPORTANCE_HIGH);
+            if(!TextUtils.isEmpty(sound)){
+                channel1.setSound(Uri.parse("android.resource://"+context.getPackageName()+"/raw/"+sound),null);
+            }
+            manager.createNotificationChannel(channel1);
+            JPushInterface.setChannel(context,channel);
+            result.success(true);
+            Log.d(TAG,"setChannelAndSound channelId="+channelId+" channel="+channel+" sound="+sound);
+
+        } catch (Throwable throwable){
+            result.success(false);
+        }
+    }
+    private void setLbsEnable(MethodCall call, Result result) {
+        HashMap<String, Object> map = call.arguments();
+        if (map == null) {
+            result.success(false);
+            return;
+        }
+        Boolean enable = (Boolean) map.get("enable");
+        if (enable == null) {
+            enable = true;
+        }
+        JPushInterface.setLbsEnable(context,enable);
+        result.success(true);
     }
 
     private void setAuth(MethodCall call, Result result){
@@ -165,6 +204,7 @@ public class JPushPlugin implements FlutterPlugin, MethodCallHandler {
     private void enableAutoWakeup(MethodCall call, Result result) {
         HashMap<String, Object> map = call.arguments();
         if (map == null) {
+            result.success(false);
             return;
         }
         Boolean enable = (Boolean) map.get("enable");
@@ -175,25 +215,7 @@ public class JPushPlugin implements FlutterPlugin, MethodCallHandler {
         result.success(true);
     }
 
-    // 主线程再返回数据
-    public void runMainThread(final Map<String, Object> map, final Result result, final String method) {
-        Log.d(TAG, "runMainThread:" + "map = " + map + ",method =" + method);
-        android.os.Handler handler = new Handler(Looper.getMainLooper());
-        handler.post(new Runnable() {
-            @Override
-            public void run() {
-                if (result == null && method != null) {
-                    if( null != channel){
-                        channel.invokeMethod(method,map);
-                    }else {
-                        Log.d(TAG,"channel is null do nothing");
-                    }
-                } else {
-                    result.success(map);
-                }
-            }
-        });
-    }
+
 
     public void setup(MethodCall call, Result result) {
         Log.d(TAG, "setup :" + call.arguments);
@@ -212,8 +234,7 @@ public class JPushPlugin implements FlutterPlugin, MethodCallHandler {
         JPushInterface.setNotificationCallBackEnable(context, true);
         String channel = (String) map.get("channel");
         JPushInterface.setChannel(context, channel);
-
-        JPushPlugin.instance.dartIsReady = true;
+        JPushHelper.getInstance().setDartIsReady(true);
 
         // try to clean getRid cache
         scheduleCache();
@@ -222,39 +243,8 @@ public class JPushPlugin implements FlutterPlugin, MethodCallHandler {
 
     public void scheduleCache() {
         Log.d(TAG, "scheduleCache:");
-        if(JPushPlugin.instance==null||JPushPlugin.instance.channel==null){
-            return;
-        }
-        List<Object> tempList = new ArrayList<Object>();
-
-        if (dartIsReady) {
-            // try to shedule notifcation cache
-            List<Map<String, Object>> openNotificationCacheList = JPushPlugin.openNotificationCache;
-            for (Map<String, Object> notification : openNotificationCacheList) {
-                JPushPlugin.instance.channel.invokeMethod("onOpenNotification", notification);
-                tempList.add(notification);
-            }
-            openNotificationCacheList.removeAll(tempList);
-        }
-
-        if (context == null) {
-            Log.d(TAG, "scheduleCache，register context is nil.");
-            return;
-        }
-
-        String rid = JPushInterface.getRegistrationID(context);
-        boolean ridAvailable = rid != null && !rid.isEmpty();
-        if (ridAvailable && dartIsReady) {
-            // try to schedule get rid cache
-            tempList.clear();
-            List<Result> resultList = JPushPlugin.instance.getRidCache;
-            for (Result res : resultList) {
-                Log.d(TAG, "scheduleCache rid = " + rid);
-                res.success(rid);
-                tempList.add(res);
-            }
-            resultList.removeAll(tempList);
-        }
+        JPushHelper.getInstance().dispatchNotification();
+        JPushHelper.getInstance().dispatchRid();
     }
 
     public void setTags(MethodCall call, Result result) {
@@ -263,7 +253,7 @@ public class JPushPlugin implements FlutterPlugin, MethodCallHandler {
         List<String> tagList = call.arguments();
         Set<String> tags = new HashSet<>(tagList);
         sequence += 1;
-        callbackMap.put(sequence, result);
+        JPushHelper.getInstance().addCallback(sequence,result);
         JPushInterface.setTags(context, sequence, tags);
     }
 
@@ -271,7 +261,7 @@ public class JPushPlugin implements FlutterPlugin, MethodCallHandler {
         Log.d(TAG, "cleanTags:");
 
         sequence += 1;
-        callbackMap.put(sequence, result);
+        JPushHelper.getInstance().addCallback(sequence,result);
         JPushInterface.cleanTags(context, sequence);
     }
 
@@ -281,7 +271,7 @@ public class JPushPlugin implements FlutterPlugin, MethodCallHandler {
         List<String> tagList = call.arguments();
         Set<String> tags = new HashSet<>(tagList);
         sequence += 1;
-        callbackMap.put(sequence, result);
+        JPushHelper.getInstance().addCallback(sequence,result);
         JPushInterface.addTags(context, sequence, tags);
     }
 
@@ -291,7 +281,7 @@ public class JPushPlugin implements FlutterPlugin, MethodCallHandler {
         List<String> tagList = call.arguments();
         Set<String> tags = new HashSet<>(tagList);
         sequence += 1;
-        callbackMap.put(sequence, result);
+        JPushHelper.getInstance().addCallback(sequence,result);
         JPushInterface.deleteTags(context, sequence, tags);
     }
 
@@ -299,14 +289,14 @@ public class JPushPlugin implements FlutterPlugin, MethodCallHandler {
         Log.d(TAG, "getAllTags： ");
 
         sequence += 1;
-        callbackMap.put(sequence, result);
+        JPushHelper.getInstance().addCallback(sequence,result);
         JPushInterface.getAllTags(context, sequence);
     }
     public void getAlias(MethodCall call, Result result) {
         Log.d(TAG, "getAlias： ");
 
         sequence += 1;
-        callbackMap.put(sequence, result);
+        JPushHelper.getInstance().addCallback(sequence,result);
         JPushInterface.getAlias(context, sequence);
     }
 
@@ -315,7 +305,7 @@ public class JPushPlugin implements FlutterPlugin, MethodCallHandler {
 
         String alias = call.arguments();
         sequence += 1;
-        callbackMap.put(sequence, result);
+        JPushHelper.getInstance().addCallback(sequence,result);
         JPushInterface.setAlias(context, sequence, alias);
     }
 
@@ -324,7 +314,7 @@ public class JPushPlugin implements FlutterPlugin, MethodCallHandler {
 
         String alias = call.arguments();
         sequence += 1;
-        callbackMap.put(sequence, result);
+        JPushHelper.getInstance().addCallback(sequence,result);
         JPushInterface.deleteAlias(context, sequence);
     }
 
@@ -375,7 +365,7 @@ public class JPushPlugin implements FlutterPlugin, MethodCallHandler {
 
         String rid = JPushInterface.getRegistrationID(context);
         if (rid == null || rid.isEmpty()) {
-            getRidCache.add(result);
+            JPushHelper.getInstance().addRid(result);
         } else {
             result.success(rid);
         }
@@ -433,7 +423,7 @@ public class JPushPlugin implements FlutterPlugin, MethodCallHandler {
         HashMap<String, Object> map = new HashMap();
         map.put("isEnabled", isEnabled == 1);
 
-        runMainThread(map, result, null);
+        JPushHelper.getInstance().runMainThread(map, result, null);
     }
 
     private void openSettingsForNotification(MethodCall call, Result result) {
@@ -463,7 +453,7 @@ public class JPushPlugin implements FlutterPlugin, MethodCallHandler {
             if (action.equals(JPushInterface.ACTION_REGISTRATION_ID)) {
                 String rId = intent.getStringExtra(JPushInterface.EXTRA_REGISTRATION_ID);
                 Log.d("JPushPlugin", "on get registration");
-                JPushPlugin.transmitReceiveRegistrationId(rId);
+                JPushHelper.getInstance().transmitReceiveRegistrationId(rId);
 
             } else if (action.equals(JPushInterface.ACTION_MESSAGE_RECEIVED)) {
                 handlingMessageReceive(intent);
@@ -480,7 +470,7 @@ public class JPushPlugin implements FlutterPlugin, MethodCallHandler {
             String msg = intent.getStringExtra(JPushInterface.EXTRA_MESSAGE);
             String title = intent.getStringExtra(JPushInterface.EXTRA_TITLE);
             Map<String, Object> extras = getNotificationExtras(intent);
-            JPushPlugin.transmitMessageReceive(msg, title, extras);
+            JPushHelper.getInstance().transmitMessageReceive(msg, title,extras);
         }
 
         private void handlingNotificationOpen(Context context, Intent intent) {
@@ -489,7 +479,7 @@ public class JPushPlugin implements FlutterPlugin, MethodCallHandler {
             String title = intent.getStringExtra(JPushInterface.EXTRA_NOTIFICATION_TITLE);
             String alert = intent.getStringExtra(JPushInterface.EXTRA_ALERT);
             Map<String, Object> extras = getNotificationExtras(intent);
-            JPushPlugin.transmitNotificationOpen(title, alert, extras);
+            JPushHelper.getInstance().transmitNotificationOpen(title, alert, extras);
         }
 
         private void handlingNotificationReceive(Context context, Intent intent) {
@@ -498,7 +488,7 @@ public class JPushPlugin implements FlutterPlugin, MethodCallHandler {
             String title = intent.getStringExtra(JPushInterface.EXTRA_NOTIFICATION_TITLE);
             String alert = intent.getStringExtra(JPushInterface.EXTRA_ALERT);
             Map<String, Object> extras = getNotificationExtras(intent);
-            JPushPlugin.transmitNotificationReceive(title, alert, extras);
+            JPushHelper.getInstance().transmitNotificationReceive(title, alert, extras);
         }
 
         private Map<String, Object> getNotificationExtras(Intent intent) {
@@ -518,156 +508,6 @@ public class JPushPlugin implements FlutterPlugin, MethodCallHandler {
     }
 
 
-    static void transmitMessageReceive(String message,String title, Map<String, Object> extras) {
-        Log.d(TAG, "transmitMessageReceive " + "message=" + message + "extras=" + extras);
 
-        if (instance == null||instance.channel==null) {
-            Log.d("JPushPlugin", "the instance is null");
-            return;
-        }
-        Map<String, Object> msg = new HashMap<>();
-        msg.put("message", message);
-        msg.put("alert", title);
-        msg.put("extras", extras);
-
-        JPushPlugin.instance.channel.invokeMethod("onReceiveMessage", msg);
-    }
-
-    static void transmitNotificationOpen(String title, String alert, Map<String, Object> extras) {
-        Log.d(TAG, "transmitNotificationOpen " + "title=" + title + "alert=" + alert + "extras=" + extras);
-
-        Map<String, Object> notification = new HashMap<>();
-        notification.put("title", title);
-        notification.put("alert", alert);
-        notification.put("extras", extras);
-        JPushPlugin.openNotificationCache.add(notification);
-
-        if (instance == null||instance.channel==null) {
-            Log.d("JPushPlugin", "the instance is null");
-            return;
-        }
-
-        if (instance.dartIsReady) {
-            Log.d("JPushPlugin", "instance.dartIsReady is true");
-            JPushPlugin.instance.channel.invokeMethod("onOpenNotification", notification);
-            JPushPlugin.openNotificationCache.remove(notification);
-        }
-
-    }
-
-    public static void onNotifyMessageUnShow( NotificationMessage notificationMessage) {
-        Log.e(TAG,"[onNotifyMessageUnShow] message:"+notificationMessage);
-        if (instance == null||instance.channel==null) {
-            Log.d("JPushPlugin", "the instance is null");
-            return;
-        }
-        Map<String, Object> notification= new HashMap<>();
-        notification.put("title", notificationMessage.notificationTitle);
-        notification.put("alert", notificationMessage.notificationContent);
-        notification.put("extras", getExtras(notificationMessage));
-        JPushPlugin.instance.channel.invokeMethod("onNotifyMessageUnShow", notification);
-    }
-    public static void onConnected( boolean isConnected) {
-        Log.e(TAG,"[onConnected] :"+isConnected);
-        if (instance == null||instance.channel==null) {
-            Log.d("JPushPlugin", "the instance is null");
-            return;
-        }
-        Map<String, Object> results= new HashMap<>();
-        results.put("result", isConnected);
-        JPushPlugin.instance.channel.invokeMethod("onConnected", results);
-    }
-    public static void onInAppMessageShow( NotificationMessage notificationMessage) {
-        Log.e(TAG,"[onInAppMessageShow] :"+notificationMessage);
-        if (instance == null||instance.channel==null) {
-            Log.d("JPushPlugin", "the instance is null");
-            return;
-        }
-        Map<String, Object> notification= new HashMap<>();
-        notification.put("title", notificationMessage.inAppMsgTitle);
-        notification.put("alert", notificationMessage.inAppMsgContentBody);
-        notification.put("messageId", notificationMessage.msgId);
-        notification.put("inAppShowTarget",  notificationMessage.inAppExtras);
-        notification.put("inAppClickAction",  notificationMessage.inAppClickAction);
-        notification.put("inAppExtras", notificationMessage.inAppExtras);
-        JPushPlugin.instance.channel.invokeMethod("onInAppMessageShow", notification);
-    }
-    public static void onInAppMessageClick( NotificationMessage notificationMessage) {
-        Log.e(TAG,"[onInAppMessageClick] :"+notificationMessage);
-        if (instance == null||instance.channel==null) {
-            Log.d("JPushPlugin", "the instance is null");
-            return;
-        }
-        Map<String, Object> notification= new HashMap<>();
-        notification.put("title", notificationMessage.inAppMsgTitle);
-        notification.put("alert", notificationMessage.inAppMsgContentBody);
-        notification.put("messageId", notificationMessage.msgId);
-        notification.put("inAppShowTarget",  notificationMessage.inAppExtras);
-        notification.put("inAppClickAction",  notificationMessage.inAppClickAction);
-        notification.put("inAppExtras", notificationMessage.inAppExtras);
-        JPushPlugin.instance.channel.invokeMethod("onInAppMessageClick", notification);
-    }
-
-
-    private static Map<String,Object> getExtras(NotificationMessage notificationMessage){
-        Map<String, Object> extras= new HashMap<>();
-        try {
-            extras.put(JPushInterface.EXTRA_MSG_ID, notificationMessage.msgId);
-            extras.put(JPushInterface.EXTRA_NOTIFICATION_ID, notificationMessage.notificationId);
-            extras.put(JPushInterface.EXTRA_ALERT_TYPE, notificationMessage.notificationAlertType + "");
-            if (!TextUtils.isEmpty(notificationMessage.notificationExtras)) {
-                extras.put(JPushInterface.EXTRA_EXTRA, notificationMessage.notificationExtras);
-            }
-            if (notificationMessage.notificationStyle == 1 && !TextUtils.isEmpty(notificationMessage.notificationBigText)) {
-                extras.put(JPushInterface.EXTRA_BIG_TEXT, notificationMessage.notificationBigText);
-            } else if (notificationMessage.notificationStyle == 2 && !TextUtils.isEmpty(notificationMessage.notificationInbox)) {
-                extras.put(JPushInterface.EXTRA_INBOX, notificationMessage.notificationInbox);
-            } else if ((notificationMessage.notificationStyle == 3) && !TextUtils.isEmpty(notificationMessage.notificationBigPicPath)) {
-                extras.put(JPushInterface.EXTRA_BIG_PIC_PATH, notificationMessage.notificationBigPicPath);
-            }
-            if (!(notificationMessage.notificationPriority == 0)) {
-                extras.put(JPushInterface.EXTRA_NOTI_PRIORITY, notificationMessage.notificationPriority + "");
-            }
-            if (!TextUtils.isEmpty(notificationMessage.notificationCategory)) {
-                extras.put(JPushInterface.EXTRA_NOTI_CATEGORY, notificationMessage.notificationCategory);
-            }
-            if (!TextUtils.isEmpty(notificationMessage.notificationSmallIcon)) {
-                extras.put(JPushInterface.EXTRA_NOTIFICATION_SMALL_ICON, notificationMessage.notificationSmallIcon);
-            }
-            if (!TextUtils.isEmpty(notificationMessage.notificationLargeIcon)) {
-                extras.put(JPushInterface.EXTRA_NOTIFICATION_LARGET_ICON, notificationMessage.notificationLargeIcon);
-            }
-        }catch (Throwable e){
-            Log.e(TAG,"[onNotifyMessageUnShow] e:"+e.getMessage());
-        }
-        return extras;
-    }
-    static void transmitNotificationReceive(String title, String alert, Map<String, Object> extras) {
-        Log.d(TAG, "transmitNotificationReceive " + "title=" + title + "alert=" + alert + "extras=" + extras);
-
-        if (instance == null||instance.channel==null) {
-            Log.d("JPushPlugin", "the instance is null");
-            return;
-        }
-
-        Map<String, Object> notification = new HashMap<>();
-        notification.put("title", title);
-        notification.put("alert", alert);
-        notification.put("extras", extras);
-        JPushPlugin.instance.channel.invokeMethod("onReceiveNotification", notification);
-    }
-
-    static void transmitReceiveRegistrationId(String rId) {
-        Log.d(TAG, "transmitReceiveRegistrationId： " + rId);
-
-        if (instance == null) {
-            return;
-        }
-
-        JPushPlugin.instance.jpushDidinit = true;
-
-        // try to clean getRid cache
-        JPushPlugin.instance.scheduleCache();
-    }
 
 }