+++ lib/jpush_flutter.dart
@@ -35,13 +35,19 @@ class JPush {
   EventHandler? _onInAppMessageClick;
   EventHandler? _onInAppMessageShow;
 
+  void _print(Object? object) {
+    if (kDebugMode) {
+      print(object);
+    }
+  }
+
   Future<void> setup({
     required String appKey,
     bool production = false,
     String channel = '',
     bool debug = false,
   }) async {
-    print(flutter_log + "setup:");
+    _print(flutter_log + "setup:");
 
     await _channel.invokeMethod('setup', {
       'appKey': appKey,
@@ -51,11 +57,25 @@ class JPush {
     });
   }
 
+  Future<void> setChannelAndSound({
+    String channel = '',
+    String channelID = '',
+    String sound = '',
+  }) async {
+    if (_platform.isIOS) {
+      return;
+    }
+    _print(flutter_log + "setChannelAndSound:");
+
+    await _channel.invokeMethod('setChannelAndSound',
+        {'channel': channel, 'channel_id': channelID, 'sound': sound});
+  }
+
   //APP活跃在前台时是否展示通知
   Future<void> setUnShowAtTheForeground({
     bool unShow = false,
   }) async {
-    print(flutter_log + "setUnShowAtTheForeground:");
+    _print(flutter_log + "setUnShowAtTheForeground:");
     await _channel.invokeMethod('setUnShowAtTheForeground', {'UnShow': unShow});
   }
 
@@ -74,8 +94,16 @@ class JPush {
   }
 
   Future<void> setAuth({bool enable = true}) async {
-    print(flutter_log + "setAuth:");
+    _print(flutter_log + "setAuth:");
     await _channel.invokeMethod('setAuth', {'enable': enable});
+  }
+
+  Future<void> setLbsEnable({bool enable = true}) async {
+    if (_platform.isIOS) {
+      return;
+    }
+    _print(flutter_log + "setLbsEnable:");
+    await _channel.invokeMethod('setLbsEnable', {'enable': enable});
   }
 
   ///
@@ -91,7 +119,7 @@ class JPush {
     EventHandler? onInAppMessageClick,
     EventHandler? onInAppMessageShow,
   }) {
-    print(flutter_log + "addEventHandler:");
+    _print(flutter_log + "addEventHandler:");
 
     _onReceiveNotification = onReceiveNotification;
     _onOpenNotification = onOpenNotification;
@@ -104,7 +132,7 @@ class JPush {
   }
 
   Future<void> _handleMethod(MethodCall call) async {
-    print(flutter_log + "_handleMethod:");
+    _print(flutter_log + "_handleMethod:");
 
     switch (call.method) {
       case "onReceiveNotification":
@@ -169,7 +197,7 @@ class JPush {
   Future<void> applyPushAuthority(
       [NotificationSettingsIOS iosSettings =
           const NotificationSettingsIOS()]) async {
-    print(flutter_log + "applyPushAuthority:");
+    _print(flutter_log + "applyPushAuthority:");
 
     if (!_platform.isIOS) {
       return;
@@ -181,7 +209,7 @@ class JPush {
   // iOS Only
   // 进入页面， pageName：页面名  请与pageLeave配套使用
   Future<void> pageEnterTo(String pageName) async {
-    print(flutter_log + "pageEnterTo:" + pageName);
+    _print(flutter_log + "pageEnterTo:" + pageName);
     if (!_platform.isIOS) {
       return;
     }
@@ -191,7 +219,7 @@ class JPush {
   // iOS Only
   // 离开页面，pageName：页面名， 请与pageEnterTo配套使用
   Future<void> pageLeave(String pageName) async {
-    print(flutter_log + "pageLeave:" + pageName);
+    _print(flutter_log + "pageLeave:" + pageName);
     if (!_platform.isIOS) {
       return;
     }
@@ -206,7 +234,7 @@ class JPush {
   /// @param {Function} fail = ({"errorCode":int}) => {  }
   ///
   Future<Map<dynamic, dynamic>?> setTags(List<String> tags) async {
-    print(flutter_log + "setTags:");
+    _print(flutter_log + "setTags:");
 
     final Map<dynamic, dynamic>? result =
         await _channel.invokeMethod('setTags', tags);
@@ -220,7 +248,7 @@ class JPush {
   /// @param {Function} fail = ({"errorCode":int}) => {  }
   ///
   Future<Map<dynamic, dynamic>?> cleanTags() async {
-    print(flutter_log + "cleanTags:");
+    _print(flutter_log + "cleanTags:");
 
     final Map<dynamic, dynamic>? result =
         await _channel.invokeMethod('cleanTags');
@@ -236,7 +264,7 @@ class JPush {
   ///
 
   Future<Map<dynamic, dynamic>?> addTags(List<String> tags) async {
-    print(flutter_log + "addTags:");
+    _print(flutter_log + "addTags:");
 
     final Map<dynamic, dynamic>? result =
         await _channel.invokeMethod('addTags', tags);
@@ -251,7 +279,7 @@ class JPush {
   /// @param {Function} fail = ({"errorCode":int}) => {  }
   ///
   Future<Map<dynamic, dynamic>?> deleteTags(List<String> tags) async {
-    print(flutter_log + "deleteTags:");
+    _print(flutter_log + "deleteTags:");
 
     final Map<dynamic, dynamic>? result =
         await _channel.invokeMethod('deleteTags', tags);
@@ -265,7 +293,7 @@ class JPush {
   /// @param {Function} fail = ({"errorCode":int}) => {  }
   ///
   Future<Map<dynamic, dynamic>?> getAllTags() async {
-    print(flutter_log + "getAllTags:");
+    _print(flutter_log + "getAllTags:");
 
     final Map<dynamic, dynamic>? result =
         await _channel.invokeMethod('getAllTags');
@@ -279,7 +307,7 @@ class JPush {
   /// @param {Function} fail = ({"errorCode":int}) => {  }
   ///
   Future<Map<dynamic, dynamic>?> getAlias() async {
-    print(flutter_log + "getAlias:");
+    _print(flutter_log + "getAlias:");
     final Map<dynamic, dynamic>? result =
         await _channel.invokeMethod('getAlias');
     return result;
@@ -294,7 +322,7 @@ class JPush {
   /// @param {Function} fail = ({"errorCode":int}) => {  }
   ///
   Future<Map<dynamic, dynamic>?> setAlias(String alias) async {
-    print(flutter_log + "setAlias:");
+    _print(flutter_log + "setAlias:");
 
     final Map<dynamic, dynamic>? result =
         await _channel.invokeMethod('setAlias', alias);
@@ -308,7 +336,7 @@ class JPush {
   /// @param {Function} fail = ({"errorCode":int}) => {  }
   ///
   Future<Map<dynamic, dynamic>?> deleteAlias() async {
-    print(flutter_log + "deleteAlias:");
+    _print(flutter_log + "deleteAlias:");
 
     final Map<dynamic, dynamic>? result =
         await _channel.invokeMethod('deleteAlias');
@@ -318,7 +346,7 @@ class JPush {
   ///
   /// 测试FCM
   Future<void> testCountryCode(String code) async {
-    print(flutter_log + "testCountryCode:" + code);
+    _print(flutter_log + "testCountryCode:" + code);
 
     await _channel.invokeMethod('testCountryCode', code);
   }
@@ -331,7 +359,7 @@ class JPush {
   /// 注意：如果是 Android 手机，目前仅支持华为手机
   ///
   Future<void> setBadge(int badge) async {
-    print(flutter_log + "setBadge:");
+    _print(flutter_log + "setBadge:");
 
     await _channel.invokeMethod('setBadge', {"badge": badge});
   }
@@ -340,7 +368,7 @@ class JPush {
   /// 停止接收推送，调用该方法后应用将不再受到推送，如果想要重新收到推送可以调用 resumePush。
   ///
   Future<void> stopPush() async {
-    print(flutter_log + "stopPush:");
+    _print(flutter_log + "stopPush:");
 
     await _channel.invokeMethod('stopPush');
   }
@@ -349,7 +377,7 @@ class JPush {
   /// 恢复推送功能。
   ///
   Future<void> resumePush() async {
-    print(flutter_log + "resumePush:");
+    _print(flutter_log + "resumePush:");
 
     await _channel.invokeMethod('resumePush');
   }
@@ -358,7 +386,7 @@ class JPush {
   /// 清空通知栏上的所有通知。
   ///
   Future<void> clearAllNotifications() async {
-    print(flutter_log + "clearAllNotifications:");
+    _print(flutter_log + "clearAllNotifications:");
 
     await _channel.invokeMethod('clearAllNotifications');
   }
@@ -368,7 +396,7 @@ class JPush {
   /// @param notificationId 通知 id，即：LocalNotification id
   ///
   Future<void> clearNotification({required int notificationId}) async {
-    print(flutter_log + "clearNotification:");
+    _print(flutter_log + "clearNotification:");
     await _channel.invokeListMethod("clearNotification", notificationId);
   }
 
@@ -380,7 +408,7 @@ class JPush {
   /// @param {Function} callback = (Object) => {}
   ///
   Future<Map<dynamic, dynamic>?> getLaunchAppNotification() async {
-    print(flutter_log + "getLaunchAppNotification:");
+    _print(flutter_log + "getLaunchAppNotification:");
 
     final Map<dynamic, dynamic>? result =
         await _channel.invokeMethod('getLaunchAppNotification');
@@ -393,7 +421,7 @@ class JPush {
   /// @param {Function} callback = (String) => {}
   ///
   Future<String> getRegistrationID() async {
-    print(flutter_log + "getRegistrationID:");
+    _print(flutter_log + "getRegistrationID:");
 
     final String rid = await _channel.invokeMethod('getRegistrationID');
     return rid;
@@ -404,7 +432,7 @@ class JPush {
   /// @param {Notification} notification
   ///
   Future<String> sendLocalNotification(LocalNotification notification) async {
-    print(flutter_log + "sendLocalNotification:");
+    _print(flutter_log + "sendLocalNotification:");
 
     await _channel.invokeMethod('sendLocalNotification', notification.toMap());